refactor: integrate WatchMode API and update movie service to support multiple sources
