+++ docker-compose.yml
@@ -43,6 +43,7 @@ services:
       MOVIE_APIS_TMDB_API_KEY: "${TMDB_API_KEY}"
       MOVIE_APIS_OMDB_API_KEY: "${OMDB_API_KEY}"
       MOVIE_APIS_SUBDL_API_KEY: "${SUBDL_API_KEY}"
+      MOVIE_APIS_WATCHMODE_API_KEY: "${WATCHMODE_API_KEY}"
       DOWNLOADS_DIRECTORY: "/app/downloads"
       STREAMING_DOWNLOAD_DIR: "/app/downloads"
       STREAMING_HLS_OUTPUT_DIR: "/app/hls_output"
