+++ server/internal/routes/movie.go
@@ -8,9 +8,10 @@ import (
 )
 
 func AddMovieRouter(movieRouter *echo.Group, movieController *controllers.MovieController) {
-	movieRouter.GET("", movieController.GetMovies)
+	movieRouter.GET("/popular", movieController.GetMovies)
 	movieRouter.GET("/search", movieController.SearchMovies)
 	movieRouter.GET("/:id", movieController.GetMovieDetails, middlewares.Authenticated, middlewares.AttachUser)
+	movieRouter.GET("/:id/:source", movieController.GetMovieDetails, middlewares.Authenticated, middlewares.AttachUser)
 }
 
 func AddCommentRouter(commentRouter *echo.Group, commentController *controllers.CommentController) {
